requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# CLIP
git+https://github.com/openai/CLIP.git

# Sentence Transformers (SBERT) - cho Track B
sentence-transformers>=2.2.0

# FAISS - Vector Search
faiss-cpu>=1.7.4

# BM25 - cho Track B
rank-bm25>=0.2.2

# UI
streamlit>=1.28.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
opencv-python>=4.8.0
